+++ src/app.module.ts
@@ -2,7 +2,6 @@ import { Module } from "@nestjs/common";
 import { APP_INTERCEPTOR } from "@nestjs/core";
 import { EventEmitterModule } from "@nestjs/event-emitter";
 import { ScheduleModule } from "@nestjs/schedule";
-
 import { DisableApiIfInterceptor } from "./interceptors/disable-api-if.interceptor";
 import { AdminModule } from "./modules/admin/admin.module";
 import { ConfigModule } from "./modules/config/config.module";
@@ -14,6 +13,7 @@ import { HealthModule } from "./modules/health/health.module";
 import { MediaModule } from "./modules/media/media.module";
 import { MetadataModule } from "./modules/metadata/metadata.module";
 import { ProgressModule } from "./modules/progresses/progress.module";
+import { SavefileModule } from "./modules/savefiles/savefile.module";
 import { UsersModule } from "./modules/users/users.module";
 
 @Module({
@@ -24,6 +24,7 @@ import { UsersModule } from "./modules/users/users.module";
     GamesModule,
     UsersModule,
     ProgressModule,
+    SavefileModule,
     MetadataModule,
     AdminModule,
     ScheduleModule.forRoot(),